Updated Calendar init to use class NYCHolidays
